requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
bcrypt
whitenoise
transaction
drf-nested-routers
dj-database-url
psycopg2-binary
#firebase-admin